Add device suffix to screens containing actual implementation
