+++ apps/mobile/src/navigation/ble-stack-navigator.ios.tsx
@@ -1,0 +1,28 @@
+import type { ParamListBase } from "@react-navigation/native";
+import { createNativeStackNavigator } from "@react-navigation/native-stack";
+import {BleDevicesListScreen} from "~/screens/ble-devices-list-screen/ble-devices-list-screen";
+import {BleConnectionScreen} from "~/screens/ble-connection-screen";
+
+export interface BleStackParamList extends ParamListBase {
+  BleDeviceList: undefined;
+  BleConnection: { deviceId: string };
+}
+
+const Stack = createNativeStackNavigator<BleStackParamList>();
+
+export function BleStackNavigator() {
+  return (
+      <Stack.Navigator id={undefined}>
+        <Stack.Screen
+          name="BleDeviceList"
+          component={BleDevicesListScreen}
+          options={{ title: "BLE Devices" }}
+        />
+        <Stack.Screen
+          name="BleConnection"
+          component={BleConnectionScreen}
+          options={{ title: "BLE Connection" }}
+        />
+      </Stack.Navigator>
+  );
+}
