--- apps/mobile/src/screens/serial-port-connection-screen.ios.tsx
@@ -1,3 +1,0 @@
-export function SerialPortConnectionScreen() {
-  return null;
-}
